Reviewed-on: https://codeberg.org/xmatt/pyfedi/pulls/6

+++ app/shared/community.py
@@ -9,7 +9,8 @@ from app.shared.upload import process_upload
 from app.shared.tasks import task_selector
 from app.user.utils import search_for_user
 from app.utils import authorise_api_user, blocked_communities, shorten_string, gibberish, markdown_to_html, \
-    instance_banned, community_membership, joined_communities, moderating_communities, is_image_url, communities_banned_from
+    instance_banned, community_membership, joined_communities, moderating_communities, is_image_url, \
+    communities_banned_from, piefed_markdown_to_lemmy_markdown
 from app.constants import *
 
 from flask import current_app, flash, render_template
@@ -339,28 +340,45 @@ def edit_community(input, community, src, auth=None, uploaded_icon_file=None, up
         return user.id
 
 
-def toggle_community_notification(community_id: int, src, auth=None):
-    # Toggle whether the current user is subscribed to notifications for activity in this community
-    if src == SRC_API:
-        user_id = authorise_api_user(auth)
-    else:
-        user_id = current_user.id
+def subscribe_community(community_id: int, subscribe, src, auth=None):
+    community = Community.query.filter_by(id=community_id, banned=False).one()
+    user_id = authorise_api_user(auth) if src == SRC_API else current_user.id
 
-    community = Community.query.filter_by(id=community_id).one()
+    if src == SRC_WEB:
+        subscribe = False if community.notify_new_posts(user_id) else True
 
-    existing_notification = NotificationSubscription.query.filter(NotificationSubscription.entity_id == community.id,
-                                                                  NotificationSubscription.user_id == user_id,
-                                                                  NotificationSubscription.type == NOTIF_COMMUNITY).first()
-    if existing_notification:
-        db.session.delete(existing_notification)
-        db.session.commit()
-    else:  # no subscription yet, so make one
-        if community.id not in communities_banned_from(user_id):
-            new_notification = NotificationSubscription(name=shorten_string(_('New posts in %(community_name)s', community_name=community.title)),
-                                                        user_id=user_id, entity_id=community.id,
-                                                        type=NOTIF_COMMUNITY)
-            db.session.add(new_notification)
+    existing_notification = NotificationSubscription.query.filter_by(entity_id=community_id, user_id=user_id,
+                                                                     type=NOTIF_COMMUNITY).first()
+    if subscribe == False:
+        if existing_notification:
+            db.session.delete(existing_notification)
             db.session.commit()
+        else:
+            msg = 'A subscription for this community did not exist.'
+            if src == SRC_API:
+                raise Exception(msg)
+            else:
+                flash(_(msg))
+
+    else:
+        if existing_notification:
+            msg = 'A subscription for this community already existed.'
+            if src == SRC_API:
+                raise Exception(msg)
+            else:
+                flash(_(msg))
+        else:
+            if community_id in communities_banned_from(user_id):
+                msg = 'You are banned from this community.'
+                if src == SRC_API:
+                    raise Exception(msg)
+                else:
+                    flash(_(msg))
+            else:
+                new_notification = NotificationSubscription(name=shorten_string(_('New posts in %(community_name)s', community_name=community.title)),
+                                                            user_id=user_id, entity_id=community_id, type=NOTIF_COMMUNITY)
+                db.session.add(new_notification)
+                db.session.commit()
 
     if src == SRC_API:
         return user_id
